feat: 전체 model 관계설정 수정 및 super 부모요소 추가 업체 서비스 부분 userId find 로직 추가

+++ 1controllers/offer.controller.js
@@ -27,7 +27,7 @@ class OfferController {
 
     const { companyId, offerId } = req.params;
 
-    const userId = res.locals.user;
+    const userId = res.locals.userId;
 
     const { status, message } = await this.offerService.updateOffer(
       offerId,
@@ -45,9 +45,9 @@ class OfferController {
   deleteOffer = async (req, res, next) => {
     const { offerId } = req.params;
 
-    // const userId = res.locals.user
+    const userId = res.locals.userId;
 
-    const { status, message } = await this.offerService.destroyOffer(offerId);
+    const { status, message } = await this.offerService.destroyOffer(offerId, userId);
     res.status(status).json(message);
   };
